requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
python-dotenv==1.0.1
transformers>=4.43.3
accelerate>=0.20.3
torch>=2.0.0
slowapi
python-multipart==0.0.6
